backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
eventlet==0.35.2
requests==2.31.0
websockets==12.0
redis==5.0.4
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
ta==0.11.0
python-binance==1.0.19
ccxt==4.3.22
python-dotenv==1.0.1
gunicorn==22.0.0
cryptography==42.0.8
hmac
hashlib
